Finding Measurements of an Object within an Image using MATLAB
In today’s Digital World, there are many digital applications that help us take pictures and 
videos of the beautiful nature. With the help of our project, we can use these images to 
measure the dimensions such as length of those objects. We use the software MATLAB to 
accomplish this certain objective. 
As technology is rapidly changing and improving due to the massive usage of various 
instruments we have designed a code with the help of MATLAB which helps in finding the 
parameters of an object by only using it image. This helps in finding any parameters(big/small) 
easier as we only the image of the machine.
MATLAB is a high-level language and interactive environment for computer computation, 
visualization, and programming. Image Processing Toolbox is an application available for use in 
MATLAB.